requirements/base.txt
django-model-utils==3.0.0
djangorestframework==3.7.3
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
sorl-thumbnail==12.4.1
